Added confirmation dialogue to delete buttons

+++ client/src/app/pages/admin/advanced/advanced.component.ts
@@ -21,7 +21,9 @@ export class AdvancedComponent implements OnInit {
   }
 
   deleteAllAppointments() {
-    this.appointmentService.deleteAllAppointments();
+    if(confirm("Are you sure you wish to delete ALL appointments for all families?")) {
+      this.appointmentService.deleteAllAppointments();
+    }
   }
 
 }
